third_party/blink/web_tests/http/tests/devtools/profiler/heap-snapshot-summary-show-ranges-expected.txt

Tests showing several node ranges in the Summary view of detailed heap snapshot.

Profiler was enabled.
Detailed heap profiles were enabled.

Running: testShowAll
Row found: true

Step 3
Retrieved ranges: [{"from":0,"to":10}]
[0] A 7730.0 k 2  0.0 k 2
[1] A 2130.0 k 2  0.0 k 2
[2] A 6130.0 k 2  0.0 k 2
[3] A 8530.0 k 2  0.0 k 2
[4] A 18130.0 k 2  0.0 k 2
[5] A 4130.0 k 2  0.0 k 2
[6] A 14130.0 k 2  0.0 k 2
[7] A 18530.0 k 2  0.0 k 2
[8] A 5730.0 k 2  0.0 k 2
[9] A 5330.0 k 2  0.0 k 2
[10] Show 10 beforeShow all 40Show 10 after

Step 4
Retrieved ranges: [{"from":0,"to":10},{"from":30,"to":40}]
[0] A 7730.0 k 2  0.0 k 2
[1] A 2130.0 k 2  0.0 k 2
[2] A 6130.0 k 2  0.0 k 2
[3] A 8530.0 k 2  0.0 k 2
[4] A 18130.0 k 2  0.0 k 2
[5] A 4130.0 k 2  0.0 k 2
[6] A 14130.0 k 2  0.0 k 2
[7] A 18530.0 k 2  0.0 k 2
[8] A 5730.0 k 2  0.0 k 2
[9] A 5330.0 k 2  0.0 k 2
[10] Show 10 beforeShow all 20Show 10 after
[11] A 13330.0 k 1  0.0 k 1
[12] A 1730.0 k 1  0.0 k 1
[13] A 14530.0 k 1  0.0 k 1
[14] A 17730.0 k 1  0.0 k 1
[15] A 19730.0 k 1  0.0 k 1
[16] A 18930.0 k 1  0.0 k 1
[17] A 2530.0 k 1  0.0 k 1
[18] A 10930.0 k 1  0.0 k 1
[19] A 20130.0 k 1  0.0 k 1
[20] A 19330.0 k 1  0.0 k 1
[21] Show 10 beforeShow all 10Show 10 after

Step 5
Retrieved ranges: [{"from":0,"to":10},{"from":20,"to":25},{"from":30,"to":40}]
[0] A 7730.0 k 2  0.0 k 2
[1] A 2130.0 k 2  0.0 k 2
[2] A 6130.0 k 2  0.0 k 2
[3] A 8530.0 k 2  0.0 k 2
[4] A 18130.0 k 2  0.0 k 2
[5] A 4130.0 k 2  0.0 k 2
[6] A 14130.0 k 2  0.0 k 2
[7] A 18530.0 k 2  0.0 k 2
[8] A 5730.0 k 2  0.0 k 2
[9] A 5330.0 k 2  0.0 k 2
[10] Show 10 beforeShow all 10Show 10 after
[11] A 12930.0 k 2  0.0 k 2
[12] A 8130.0 k 2  0.0 k 2
[13] A 15730.0 k 2  0.0 k 2
[14] A 13730.0 k 1  0.0 k 1
[15] A 16930.0 k 1  0.0 k 1
[16] Show 10 beforeShow all 5Show 10 after
[17] A 13330.0 k 1  0.0 k 1
[18] A 1730.0 k 1  0.0 k 1
[19] A 14530.0 k 1  0.0 k 1
[20] A 17730.0 k 1  0.0 k 1
[21] A 19730.0 k 1  0.0 k 1
[22] A 18930.0 k 1  0.0 k 1
[23] A 2530.0 k 1  0.0 k 1
[24] A 10930.0 k 1  0.0 k 1
[25] A 20130.0 k 1  0.0 k 1
[26] A 19330.0 k 1  0.0 k 1
[27] Show 10 beforeShow all 10Show 10 after

Step 6
Retrieved ranges: [{"from":0,"to":10},{"from":20,"to":25},{"from":28,"to":40}]
[0] A 7730.0 k 2  0.0 k 2
[1] A 2130.0 k 2  0.0 k 2
[2] A 6130.0 k 2  0.0 k 2
[3] A 8530.0 k 2  0.0 k 2
[4] A 18130.0 k 2  0.0 k 2
[5] A 4130.0 k 2  0.0 k 2
[6] A 14130.0 k 2  0.0 k 2
[7] A 18530.0 k 2  0.0 k 2
[8] A 5730.0 k 2  0.0 k 2
[9] A 5330.0 k 2  0.0 k 2
[10] Show 10 beforeShow all 10Show 10 after
[11] A 12930.0 k 2  0.0 k 2
[12] A 8130.0 k 2  0.0 k 2
[13] A 15730.0 k 2  0.0 k 2
[14] A 13730.0 k 1  0.0 k 1
[15] A 16930.0 k 1  0.0 k 1
[16] Show 10 beforeShow all 3Show 10 after
[17] A 4530.0 k 1  0.0 k 1
[18] A 3330.0 k 1  0.0 k 1
[19] A 13330.0 k 1  0.0 k 1
[20] A 1730.0 k 1  0.0 k 1
[21] A 14530.0 k 1  0.0 k 1
[22] A 17730.0 k 1  0.0 k 1
[23] A 19730.0 k 1  0.0 k 1
[24] A 18930.0 k 1  0.0 k 1
[25] A 2530.0 k 1  0.0 k 1
[26] A 10930.0 k 1  0.0 k 1
[27] A 20130.0 k 1  0.0 k 1
[28] A 19330.0 k 1  0.0 k 1
[29] Show 10 beforeShow all 10Show 10 after

Step 7
Retrieved ranges: [{"from":0,"to":10},{"from":18,"to":26},{"from":28,"to":40}]
[0] A 7730.0 k 2  0.0 k 2
[1] A 2130.0 k 2  0.0 k 2
[2] A 6130.0 k 2  0.0 k 2
[3] A 8530.0 k 2  0.0 k 2
[4] A 18130.0 k 2  0.0 k 2
[5] A 4130.0 k 2  0.0 k 2
[6] A 14130.0 k 2  0.0 k 2
[7] A 18530.0 k 2  0.0 k 2
[8] A 5730.0 k 2  0.0 k 2
[9] A 5330.0 k 2  0.0 k 2
[10] Show 10 beforeShow all 8Show 10 after
[11] A 12530.0 k 2  0.0 k 2
[12] A 11730.0 k 2  0.0 k 2
[13] A 12930.0 k 2  0.0 k 2
[14] A 8130.0 k 2  0.0 k 2
[15] A 15730.0 k 2  0.0 k 2
[16] A 13730.0 k 1  0.0 k 1
[17] A 16930.0 k 1  0.0 k 1
[18] A 20530.0 k 1  0.0 k 1
[19] Show 10 beforeShow all 2Show 10 after
[20] A 4530.0 k 1  0.0 k 1
[21] A 3330.0 k 1  0.0 k 1
[22] A 13330.0 k 1  0.0 k 1
[23] A 1730.0 k 1  0.0 k 1
[24] A 14530.0 k 1  0.0 k 1
[25] A 17730.0 k 1  0.0 k 1
[26] A 19730.0 k 1  0.0 k 1
[27] A 18930.0 k 1  0.0 k 1
[28] A 2530.0 k 1  0.0 k 1
[29] A 10930.0 k 1  0.0 k 1
[30] A 20130.0 k 1  0.0 k 1
[31] A 19330.0 k 1  0.0 k 1
[32] Show 10 beforeShow all 10Show 10 after

Step 8
Retrieved ranges: [{"from":0,"to":10},{"from":15,"to":45}]
[0] A 7730.0 k 2  0.0 k 2
[1] A 2130.0 k 2  0.0 k 2
[2] A 6130.0 k 2  0.0 k 2
[3] A 8530.0 k 2  0.0 k 2
[4] A 18130.0 k 2  0.0 k 2
[5] A 4130.0 k 2  0.0 k 2
[6] A 14130.0 k 2  0.0 k 2
[7] A 18530.0 k 2  0.0 k 2
[8] A 5730.0 k 2  0.0 k 2
[9] A 5330.0 k 2  0.0 k 2
[10] Show 10 beforeShow all 5Show 10 after
[11] A 6530.0 k 2  0.0 k 2
[12] A 11330.0 k 2  0.0 k 2
[13] A 15330.0 k 2  0.0 k 2
[14] A 12530.0 k 2  0.0 k 2
[15] A 11730.0 k 2  0.0 k 2
[16] A 12930.0 k 2  0.0 k 2
[17] A 8130.0 k 2  0.0 k 2
[18] A 15730.0 k 2  0.0 k 2
[19] A 13730.0 k 1  0.0 k 1
[20] A 16930.0 k 1  0.0 k 1
[21] A 20530.0 k 1  0.0 k 1
[22] A 3730.0 k 1  0.0 k 1
[23] A 4930.0 k 1  0.0 k 1
[24] A 4530.0 k 1  0.0 k 1
[25] A 3330.0 k 1  0.0 k 1
[26] A 13330.0 k 1  0.0 k 1
[27] A 1730.0 k 1  0.0 k 1
[28] A 14530.0 k 1  0.0 k 1
[29] A 17730.0 k 1  0.0 k 1
[30] A 19730.0 k 1  0.0 k 1
[31] A 18930.0 k 1  0.0 k 1
[32] A 2530.0 k 1  0.0 k 1
[33] A 10930.0 k 1  0.0 k 1
[34] A 20130.0 k 1  0.0 k 1
[35] A 19330.0 k 1  0.0 k 1
[36] A 14930.0 k 1  0.0 k 1
[37] A 6930.0 k 1  0.0 k 1
[38] A 2930.0 k 0  0.0 k 0
[39] A 12130.0 k 0  0.0 k 0
[40] A 16530.0 k 0  0.0 k 0
[41] Show 10 beforeShow all 5Show 10 after

Step 9
Retrieved ranges: [{"from":0,"to":45}]
[0] A 7730.0 k 2  0.0 k 2
[1] A 2130.0 k 2  0.0 k 2
[2] A 6130.0 k 2  0.0 k 2
[3] A 8530.0 k 2  0.0 k 2
[4] A 18130.0 k 2  0.0 k 2
[5] A 4130.0 k 2  0.0 k 2
[6] A 14130.0 k 2  0.0 k 2
[7] A 18530.0 k 2  0.0 k 2
[8] A 5730.0 k 2  0.0 k 2
[9] A 5330.0 k 2  0.0 k 2
[10] A 8930.0 k 2  0.0 k 2
[11] A 9730.0 k 2  0.0 k 2
[12] A 7330.0 k 2  0.0 k 2
[13] A 10130.0 k 2  0.0 k 2
[14] A 1330.0 k 2  0.0 k 2
[15] A 6530.0 k 2  0.0 k 2
[16] A 11330.0 k 2  0.0 k 2
[17] A 15330.0 k 2  0.0 k 2
[18] A 12530.0 k 2  0.0 k 2
[19] A 11730.0 k 2  0.0 k 2
[20] A 12930.0 k 2  0.0 k 2
[21] A 8130.0 k 2  0.0 k 2
[22] A 15730.0 k 2  0.0 k 2
[23] A 13730.0 k 1  0.0 k 1
[24] A 16930.0 k 1  0.0 k 1
[25] A 20530.0 k 1  0.0 k 1
[26] A 3730.0 k 1  0.0 k 1
[27] A 4930.0 k 1  0.0 k 1
[28] A 4530.0 k 1  0.0 k 1
[29] A 3330.0 k 1  0.0 k 1
[30] A 13330.0 k 1  0.0 k 1
[31] A 1730.0 k 1  0.0 k 1
[32] A 14530.0 k 1  0.0 k 1
[33] A 17730.0 k 1  0.0 k 1
[34] A 19730.0 k 1  0.0 k 1
[35] A 18930.0 k 1  0.0 k 1
[36] A 2530.0 k 1  0.0 k 1
[37] A 10930.0 k 1  0.0 k 1
[38] A 20130.0 k 1  0.0 k 1
[39] A 19330.0 k 1  0.0 k 1
[40] A 14930.0 k 1  0.0 k 1
[41] A 6930.0 k 1  0.0 k 1
[42] A 2930.0 k 0  0.0 k 0
[43] A 12130.0 k 0  0.0 k 0
[44] A 16530.0 k 0  0.0 k 0
[45] Show 10 beforeShow all 5Show 10 after

Profiler was disabled.
